Refactor: move find_cycle_ends function to helpers
